requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
albumentations==1.1.0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
compay-syntax==0.4.0
cutadapt==3.0
cycler==0.10.0
decorator==4.4.2
dnaio==0.4.4
efficientnet-pytorch==0.7.0
einops==0.3.0
fsspec==2021.6.1
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
imageio==2.9.0
ipython-genutils==0.2.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
networkx==2.5
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.5.4.58
openslide-python==1.1.2
packaging==20.9
pandas==1.2.3
Pillow==8.4.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.5.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
qudida==0.0.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
six=1.16.0
smmap==3.0.5
syntax==0.1.3.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.3.17
torch==1.8.1
torchmetrics==0.6.2
torchvision==0.9.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typeguard==2.11.1
typing-extensions=4.0.1
urllib3==1.26.4
Werkzeug==1.0.1
xmltodict==0.12.0
xopen==1.0.1
yarl==1.6.3
